feat: ensure product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -4,14 +4,14 @@ import { eq } from "drizzle-orm";
 import { headers } from "next/headers";
 
 import { db } from "@/db";
-import { cartItemTable, cartTable } from "@/db/schema";
+import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 import { z } from "zod";
 
-export const removeProductFromCart = async (data: z.infer<typeof removeProductFromCartSchema>) => {
-  removeProductFromCartSchema.parse(data);
+export const decreaseCartProductQuantity = async (data: z.infer<typeof decreaseCartProductQuantitySchema>) => {
+  decreaseCartProductQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
@@ -33,6 +33,12 @@ export const removeProductFromCart = async (data: z.infer<typeof removeProductFr
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
-  
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+  await db
+  .update(cartItemTable)
+  .set({quantity: cartItem.quantity - 1,})
+  .where(eq(cartItemTable.id, cartItem.id));
 };
